Logout function fixed (now redirect to welcome page). Navbar.js created to clear App.js
