+++ client/src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import {BrowserRouter as Router, Routes, Route, Link, useNavigate} from 'react-router-dom'
+import {BrowserRouter as Router, Routes, Route, Link} from 'react-router-dom'
 import Login from "./pages/Login";
 import Home from "./pages/Home";
 import { AuthContext } from "./helpers/AuthContext"
@@ -16,6 +16,9 @@ import AuxiliaresDisponibles from "./pages/AuxiliaresDisponibles";
 import EditPersonaDependiente from "./pages/EditPersonaDependiente";
 import EditAuxiliar from "./pages/EditAuxiliar";
 import PersonasDependientesAsignadas from "./pages/PersonasDependientesAsignadas";
+import Welcome from "./pages/Welcome";
+import { useNavigate } from "react-router-dom";
+import Navbar from "./pages/Navbar";
 
 
 function App() {
@@ -49,48 +52,14 @@ function App() {
             });
     }, []);
 
-    const logout = () => {
-        localStorage.removeItem("accessToken");
-        setAuthState({
-            username: "",
-            id: 0,
-            rol: "",
-            status: false,
-        })
-    }
-
   return (
 
       <div className="App">
           <AuthContext.Provider value={{ authState, setAuthState }}>
               <Router>
-                  <div className="navbar">
-                      <label><Link to='/home'>Home</Link></label>
-                      {authState.rol === "COORDINADOR" &&
-                          <>
-                              <label><Link to='/coordinador/personasDependientes'>Gestionar personas dependientes</Link></label>
-                              <label><Link to='/coordinador/auxiliares'>Gestionar auxiliares</Link></label>
-                          </>
-                      }
-
-                      {authState.rol === "AUXILIAR" &&
-                      <>
-                          <label><Link to='/auxiliar/asignadas'>Mis personas asignadas</Link></label>
-                      </>
-                      }
-                      {!authState.status ? (
-                          <>
-                              <label><Link to='/login'>Login</Link></label>
-                          </>
-                      ) : (
-                          <button onClick={logout}>Logout</button>
-                      )}
-                      <div className="loggedInContainer">
-                          <h1>{authState.username}</h1>
-                      </div>
-
-                  </div>
+                  <Navbar/>
                   <Routes>
+                      <Route path='/' element={<Welcome/>}/>
                       <Route path='/home' element={<Home/>} />
                       <Route path='/login' element={<Login/>} />
                       <Route path='/coordinador/personasDependientes' element={<GestionarPersonasDependientes/>} />
