+++ client/src/pages/Welcome.js
@@ -1,0 +1,11 @@
+const Welcome = () => {
+    return(
+        <div>
+            <h1>
+                This is the welcome page
+            </h1>
+        </div>
+    )
+}
+
+export default Welcome;
